+++ plugins/vps_manager/__init__.py
@@ -1,0 +1,2 @@
+"""Ansispire VPS manager plugin."""
+